Changes in path
Made a few changes in the filesystem to divide every endpoint and method. The filenames to better reflect what each is for, divided every method into it's separated file, called the same way as the method, and inside the endpoint folder they belong to. Also switched all HTTP requests to using the APIRequests class in utilities.php