Changes in error codes
With the switch to using Requests for PHP as the HTTP requests library the API can now analyze what exactly went wrong with the outgoing requests and properly reflect so, meaning there are new and more specific error codes. A new key that was added for IUsers/GetUserInfo (steamid64_str) has been documented. New error codes for IGiveaways/GetUserInfo that tell whether the API hoster was blacklisted as well as any other limitation.